requirements.txt
Flask==2.2.5
tensorflow-cpu==2.11.0
Pillow==9.5.0
numpy==1.25.0
gunicorn
